v. 2 mediekompasset/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
requests==2.26.0
beautifulsoup4==4.9.3
spacy==3.1.3
pandas==1.3.3
scikit-learn==0.24.2
nltk==3.6.3
psycopg2-binary==2.9.1
python-dotenv==0.19.0
aiofiles==0.7.0
jinja2==3.0.1
python-multipart==0.0.5
plotly==5.3.1
dash==2.0.0
tensorflow==2.7.0
transformers==4.12.5
selenium==4.1.0
newspaper3k==0.2.8
wordcloud==1.8.1
matplotlib==3.4.3
seaborn==0.11.2
